openAPI doc updated
