+++ .eslintrc.js
@@ -20,9 +20,5 @@ module.exports = {
     '@typescript-eslint/explicit-function-return-type': 'off',
     '@typescript-eslint/explicit-module-boundary-types': 'off',
     '@typescript-eslint/no-explicit-any': 'off',
-
-    '@typescript-eslint/no-unused-vars': 'off',
-    "prettier/prettier": ["error", { "endOfLine": "auto" }]  
-
   },
 };
