+++ src/api/app.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { DbService } from './utils/db.service';
-
+import { DbService } from '../utils/db.service';
+import { Utils } from '../utils/utils.service';
 import { UserController } from './user/user.controller';
 import { UserService } from './user/user.service';
 import { ArtistController } from './artist/artist.controller';
@@ -11,7 +11,6 @@ import { AlbumController } from './album/album.controller';
 import { AlbumService } from './album/album.service';
 import { FavoritesController } from './favorites/favorites.controller';
 import { FavoritesService } from './favorites/favorites.service';
-import { Utils } from './utils/utils.service';
 
 @Module({
   imports: [],
