+++ src/api/artist/artist.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { ArtistDto } from './dto/artist.dto';
-import { DbService } from '../utils/db.service';
+import { DbService } from '../../utils/db.service';
 import { Utils } from 'src/utils/utils.service';
 import { randomUUID } from 'crypto';
