+++ src/api/artist/dto/artist.dto.ts
